server_backup/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Django==5.1.4
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pyserial==3.5
sqlparse==0.5.3
Werkzeug==3.1.3
